Decouple various queries from the Device
This commit creates an abstraction for the Device, where queries are
executed by passing a class to the Device.get() method. The class must
implement a static method .get(device), which can return an object of
that class.

This way Device does not need to know anything about how to query for
certain data. This responsibility stays entirely within the class that
implements that functionality.