Fix some null handling issues with QoS
The conversion to Pydantic JSON serialization seems to have introduced
a bug, where null fields are introduced for fields that weren't
used. This doesn't cause issues if the fields are populated, but
passing null in some fields causes 422 errors with the API.

This more strictly uses exclude_none and moves some field population
into if statements to avoid null fields.